+++ src/jsMain/kotlin/at/posselt/pfrpg2e/utils/Time.kt
@@ -58,15 +58,10 @@ fun LocalTime.toDateInputString(): String =
 fun LocalTime.isDay(): Boolean =
     hour >= 6 && hour < 18
 
-fun formatSeconds(seconds: Int, isNegative: Boolean = false): String {
+fun formatSeconds(seconds: Int): String {
     val hours = seconds / 3600
     val minutes = (seconds - hours * 3600) / 60
-    val time = "${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}"
-    return if (isNegative) {
-        "-$time"
-    } else {
-        time
-    }
+    return "${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}"
 }
 
 val GameTime.worldTimeSeconds: Int