+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.10.0] - 2026-01-26
+
+### Changed
+
+* Instead of tracking a fixed amount of travel time with a special button, the camping sheet now allows you to toggle a checkbox to determine if time passed should be stored as travel time
+
+### Fixed
+
+* Negative travel time does not show a double minus sign anymore
+
+
 ## [5.9.4] - 2026-01-25
 
 ### Fixed